pyproject.toml
[project]
name = "ChatskisApp"
version = "0.1.1"
description = "A Serviceless Mulitmodal AI ChatBot"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-core>=0.3.21",
    "langchain-groq>=0.2.1",
    "streamlit>=1.40.2",
    "google-cloud-texttospeech==2.22.0",
    "google-cloud-speech==2.29.0",
    "google-cloud-translate==3.19.0",
    "python-dotenv==1.0.1",
]
